Charles' Suggested Changes
For presentation purposes:
Changed the function: "Heal" to "HealPlayer"
Changed the function: "getHealth" to "GetHealth"
Changed the function: "RoomContents" to "GetRoomContents"
Changed the function: "attackDmg" to "GetAttackDmg"
Changed the function: "InventoryContents" to "GetInventoryContents"
Changed other function names to follow the Pascal Case format.

Using "Get" in these function names is nice as it shows that the function will display them, e.g: "GetHealth" displays the current health of the player, etc...

For presentation purposes:
Changed: "a" to "attack"
Changed: "h" to "heal"
Changed: "i" to "inv" (inventory)
Changed: "p" to "pick" (pickup item)

Although this makes typing the actions a longer process, it seems a worthwhile edit for presentation purposes. All of the descriptions in the interface are now neater, with a brief description of their purposes for the player.

Pascal Case is now used for every function name.
These changes were made for readability purposes, as well as to make the function names more consistent and clear.

Description attacking the minotaur has been edited, making it feel more appropriate.
Description for defeating the minotaur has been edited, making it feel more appropriate.
Minotaur HP has increased to 80, making it feel more challenging, as the player has 100HP.

Added a new function that causes the player to receive damage from the minotaur, making the heal function actually have a use.

After reading the reviews suggested by Charles, I acted upon the feedback and a lot of it was very helpful advice. I really liked the idea of changing the function names to begin with "Get" if they are to display their purposes on the game screen. It makes the code feel a lot tidier, complete and enhances readability. I added a function that causes the monster to attack the player to add a level of difficulty to the game.