Charles' and Ben's Suggestions and some other tweaks
After acting upon Charles' reviews and suggestions, I focused on Ben's. I really liked the idea of adding the ToLower() and Trim() functions for user input. This allowed the game to feel more complete and more playable for users. I also added comments throughout the code, for presentation purposes, as well as allowing it to be more understandable as a whole.